fix warning

+++ src/background-filter.cpp
@@ -433,10 +433,10 @@ static void processImageForBackground(struct background_removal_filter *tf,
 // Blend
 void blend_images_with_mask(cv::Mat &dst, const cv::Mat &src, const cv::Mat &mask)
 {
-  for (int i = 0; i < dst.total(); i++) {
-    const auto &srcPixel = src.at<cv::Vec4b>(i);
-    const auto maskPixel = mask.at<uchar>(i) / 255.0f;
-    auto &dstPixel = dst.at<cv::Vec4b>(i);
+  for (size_t i = 0; i < dst.total(); i++) {
+    const auto &srcPixel = src.at<cv::Vec4b>((int)i);
+    const auto maskPixel = mask.at<uchar>((int)i) / 255.0f;
+    auto &dstPixel = dst.at<cv::Vec4b>((int)i);
 
     if (maskPixel == 1.0f) {
       dstPixel = srcPixel;
